Added some collisions on fireball, main character... AND A F***ING BOXgit statusgit status!
