+++ Assets/C# Scripts/Ennemies.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: b7753b7714d8ab24aa80f0ccd44a4055
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
