+++ Assets/C# Scripts/Ennemies/Ennemies.cs
@@ -1,0 +1,16 @@
+﻿using System.Collections;
+using System.Collections.Generic;
+using UnityEngine;
+
+public class Ennemies : MonoBehaviour
+{
+    public float health = 3f;
+    // Update is called once per frame
+    void Update()
+    {
+        if (health <= 0f)
+        {
+            Destroy(gameObject);
+        }
+    }
+}
