+++ Assets/C# Scripts/Ennemies/Ennemies.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 139be079c292f074a8f24a83e133f2e7
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
