+++ Assets/GameObjects/Character.prefab
@@ -12,9 +12,11 @@ GameObject:
   - component: {fileID: 1535557953062949239}
   - component: {fileID: 1535557953062949232}
   - component: {fileID: 1535557953062949237}
+  - component: {fileID: 3782361371751249401}
+  - component: {fileID: 526024250576593454}
   m_Layer: 0
   m_Name: Character
-  m_TagString: Untagged
+  m_TagString: Player
   m_Icon: {fileID: 0}
   m_NavMeshLayer: 0
   m_StaticEditorFlags: 0
@@ -119,3 +121,41 @@ MonoBehaviour:
   FireballOffset: {x: 0.4, y: -0.4}
   Yincrement: 5
   Xincrement: 2
+  health: 3
+--- !u!58 &3782361371751249401
+CircleCollider2D:
+  m_ObjectHideFlags: 0
+  m_CorrespondingSourceObject: {fileID: 0}
+  m_PrefabInstance: {fileID: 0}
+  m_PrefabAsset: {fileID: 0}
+  m_GameObject: {fileID: 1535557953062949233}
+  m_Enabled: 1
+  m_Density: 1
+  m_Material: {fileID: 0}
+  m_IsTrigger: 0
+  m_UsedByEffector: 0
+  m_UsedByComposite: 0
+  m_Offset: {x: 0.36474767, y: -0.09379223}
+  serializedVersion: 2
+  m_Radius: 1.9396626
+--- !u!50 &526024250576593454
+Rigidbody2D:
+  serializedVersion: 4
+  m_ObjectHideFlags: 0
+  m_CorrespondingSourceObject: {fileID: 0}
+  m_PrefabInstance: {fileID: 0}
+  m_PrefabAsset: {fileID: 0}
+  m_GameObject: {fileID: 1535557953062949233}
+  m_BodyType: 1
+  m_Simulated: 1
+  m_UseFullKinematicContacts: 0
+  m_UseAutoMass: 0
+  m_Mass: 1
+  m_LinearDrag: 0
+  m_AngularDrag: 0.05
+  m_GravityScale: 1
+  m_Material: {fileID: 0}
+  m_Interpolate: 0
+  m_SleepingMode: 1
+  m_CollisionDetection: 0
+  m_Constraints: 0
