+++ Assets/GameObjects/Crate.prefab.meta
@@ -1,0 +1,7 @@
+fileFormatVersion: 2
+guid: 5a6d34c4d04126348bb9260d64b3e0c4
+PrefabImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
